Created basic getBestSellers method that prints to console
Used .filter() to parse through the BOOK class in order to select only the best sellers. It currently prints to console but the method will need to be rewritten in order to print to the cards on the Best Seller page. Also need to figure out how to  create multiple cards so the best sellers will show up. Will need the same functionality for new arrivals.  Might be better to make a new component called bookCards so that I can reuse the same code for the two different webpages